+++ packages/common-solana/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Util function to get discriminator from name (#25)
 
 ## [1.1.1] - 2025-09-10
 ### Fixed